requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
kagglehub>=0.3.0
tqdm>=4.66.0
transformers>=4.53.0
accelerate>=0.34.0
datasets>=2.21.0
peft>=0.12.0
timm>=1.0.8
sentencepiece>=0.2.0
trl>=0.10.1
